fix: improve cactus API ergonomics and consistency
- Wrap complete_stream return type in CompletionStream struct with
  Stream impl, cancel() method, and Drop-based cleanup
- Wrap transcribe_stream return type in TranscriptionSession struct
  with Stream impl, audio_tx()/cancel() accessors, and Drop cleanup
- Unify token count types: CompletionResult u32->u64, StreamResult f64->u64
  (now consistent with TranscriptionResult which already used u64)
- Add tracing::warn when recovering from poisoned inference mutex
- Update callers in llm-cactus and transcribe-cactus

+++ crates/cactus/src/stt/transcriber.rs
@@ -89,11 +89,11 @@ pub struct StreamResult {
     #[serde(default)]
     pub ram_usage_mb: f64,
     #[serde(default)]
-    pub prefill_tokens: f64,
+    pub prefill_tokens: u64,
     #[serde(default)]
-    pub decode_tokens: f64,
+    pub decode_tokens: u64,
     #[serde(default)]
-    pub total_tokens: f64,
+    pub total_tokens: u64,
 }
 
 impl std::str::FromStr for StreamResult {
